Bail from Phil-opp's magic-heavy approach
I originally followed https://os.phil-opp.com/ as the starting point,
but I was not happy about how much it abstracts things like assembly
code and the process of setting up GDT and such. It also cost me a bunch
of time. So I'm switching over to generally following intermezzOS'
version 1 of getting started, which actually walks through setting up
the necessary toolchain, along with things like Long Mode and dealing
with GRUB, even if not in great detail.